+++ tests/components/ourgroceries/test_todo.py
@@ -1,13 +1,20 @@
 """Unit tests for the OurGroceries todo platform."""
+from asyncio import TimeoutError as AsyncIOTimeoutError
 from unittest.mock import AsyncMock
 
+from aiohttp import ClientError
+from freezegun.api import FrozenDateTimeFactory
 import pytest
 
+from homeassistant.components.ourgroceries.coordinator import SCAN_INTERVAL
 from homeassistant.components.todo import DOMAIN as TODO_DOMAIN
+from homeassistant.const import STATE_UNAVAILABLE
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_component import async_update_entity
 
 from . import items_to_shopping_list
+
+from tests.common import async_fire_time_changed
 
 
 @pytest.mark.parametrize(
@@ -207,3 +214,30 @@ async def test_remove_todo_item(
     state = hass.states.get("todo.test_list")
     assert state
     assert state.state == "0"
+
+
+@pytest.mark.parametrize(
+    ("exception"),
+    [
+        (ClientError),
+        (AsyncIOTimeoutError),
+    ],
+)
+async def test_coordinator_error(
+    hass: HomeAssistant,
+    freezer: FrozenDateTimeFactory,
+    setup_integration: None,
+    ourgroceries: AsyncMock,
+    exception: Exception,
+) -> None:
+    """Test error on coordinator update."""
+    state = hass.states.get("todo.test_list")
+    assert state.state == "0"
+
+    ourgroceries.get_list_items.side_effect = exception
+    freezer.tick(SCAN_INTERVAL)
+    async_fire_time_changed(hass)
+    await hass.async_block_till_done()
+
+    state = hass.states.get("todo.test_list")
+    assert state.state == STATE_UNAVAILABLE
